feat: Business onramp throws unauthorized payment error if transaction id does not exist
I'm checking if the transaction id exists in the transactions table. I've placed a FOR share row level lock as this check is done to avoid a case where a concurrent transaction modifies the row as the check is being done. If the SELECT returns no rows, i.e the transaction does not exist, then an unauthorized payment error is thrown