fix: Storing balance without considering decimals
I was storing balance of tokens from db as a bigint, the issue with bigint is that you can't represent decimals. I've decided to then change the code that updates treasury balance to store the raw balance from the blockchain without dividing by decimals, and then I also update the decimals. Then when I get the cached balance I can divide the balance with decimals to get a number representation that has 2 decimal places.