feat: Storing hash of private key
The AES-256-GCM encryption algorithm encrypts text differently every time it is called. This is an incredible way to encrypt but it makes my previous plan of comparing plain text private key to cipher text in db invalid. I've decided to now also store a hash of the private key, then I'll later be able to compare this hash to the plain text. So when creating a new environment or rotating keys I also store the hash